refactor: 레코드 분리(#88)
